// apps/web/generated/api/models/formFieldType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type FormFieldType = typeof FormFieldType[keyof typeof FormFieldType];


export const FormFieldType = {
  text: 'text',
  textarea: 'textarea',
  number: 'number',
  decimal: 'decimal',
  boolean: 'boolean',
  date: 'date',
  datetime: 'datetime',
  email: 'email',
  phone: 'phone',
  url: 'url',
  select: 'select',
  multi_select: 'multi_select',
  file: 'file',
  json: 'json',
} as const;